Refactor imports in commerce-related modules to consolidate rootLink and commerce constants. This change improves code clarity and consistency across various components by importing multiple constants from a single source. Additionally, new utility functions for preloading files and handling commerce-specific page types have been added to enhance performance and maintainability.
