+++ blocks/commerce-account-sidebar/commerce-account-sidebar.js
@@ -1,8 +1,7 @@
 import { Icon, provider as UI } from '@dropins/tools/components.js';
 import { render as accountRenderer } from '@dropins/storefront-account/render.js';
 import { loadFragment } from '../fragment/fragment.js';
-import { CUSTOMER_ORDERS_PATH } from '../../scripts/commerce.js';
-import { rootLink } from '../../scripts/scripts.js';
+import { CUSTOMER_ORDERS_PATH, rootLink } from '../../scripts/commerce.js';
 
 export default async function decorate(block) {
   const fragment = await loadFragment('/customer/sidebar-fragment');
