+++ blocks/commerce-addresses/commerce-addresses.js
@@ -1,8 +1,12 @@
 import { Addresses } from '@dropins/storefront-account/containers/Addresses.js';
 import { render as accountRenderer } from '@dropins/storefront-account/render.js';
 import { readBlockConfig } from '../../scripts/aem.js';
-import { CUSTOMER_ADDRESS_PATH, CUSTOMER_LOGIN_PATH, checkIsAuthenticated } from '../../scripts/commerce.js';
-import { rootLink } from '../../scripts/scripts.js';
+import {
+  CUSTOMER_ADDRESS_PATH,
+  CUSTOMER_LOGIN_PATH,
+  checkIsAuthenticated,
+  rootLink,
+} from '../../scripts/commerce.js';
 
 // Initialize
 import '../../scripts/initializers/account.js';
