+++ blocks/commerce-checkout/commerce-checkout.js
@@ -76,8 +76,12 @@ import {
   setAddressOnCart,
 } from './utils.js';
 
-import { authPrivacyPolicyConsentSlot, SUPPORT_PATH, fetchPlaceholders } from '../../scripts/commerce.js';
-import { rootLink } from '../../scripts/scripts.js';
+import {
+  authPrivacyPolicyConsentSlot,
+  SUPPORT_PATH,
+  fetchPlaceholders,
+  rootLink,
+} from '../../scripts/commerce.js';
 
 // Initializers
 import '../../scripts/initializers/account.js';
