+++ blocks/commerce-order-returns/commerce-order-returns.js
@@ -6,11 +6,11 @@ import {
   RETURN_DETAILS_PATH,
   UPS_TRACKING_URL,
   checkIsAuthenticated,
+  rootLink,
 } from '../../scripts/commerce.js';
 
 // Initialize
 import '../../scripts/initializers/order.js';
-import { rootLink } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
   const isAuthenticated = checkIsAuthenticated();
