+++ blocks/commerce-returns-list/commerce-returns-list.js
@@ -2,11 +2,11 @@ import { render as orderRenderer } from '@dropins/storefront-order/render.js';
 import ReturnsList from '@dropins/storefront-order/containers/ReturnsList.js';
 import { tryRenderAemAssetsImage } from '@dropins/tools/lib/aem/assets.js';
 import { readBlockConfig } from '../../scripts/aem.js';
-import { rootLink } from '../../scripts/scripts.js';
 import {
+  rootLink,
   CUSTOMER_LOGIN_PATH,
-  CUSTOMER_RETURN_DETAILS_PATH,
   CUSTOMER_ORDER_DETAILS_PATH,
+  CUSTOMER_RETURN_DETAILS_PATH,
   CUSTOMER_RETURNS_PATH,
   UPS_TRACKING_URL,
   checkIsAuthenticated,
