+++ blocks/commerce-search-order/commerce-search-order.js
@@ -3,12 +3,16 @@ import { OrderSearch } from '@dropins/storefront-order/containers/OrderSearch.js
 import { render as authRenderer } from '@dropins/storefront-auth/render.js';
 import { render as orderRenderer } from '@dropins/storefront-order/render.js';
 import { events } from '@dropins/tools/event-bus.js';
-import { CUSTOMER_ORDER_DETAILS_PATH, ORDER_DETAILS_PATH, checkIsAuthenticated } from '../../scripts/commerce.js';
+import {
+  CUSTOMER_ORDER_DETAILS_PATH,
+  ORDER_DETAILS_PATH,
+  checkIsAuthenticated,
+  rootLink,
+} from '../../scripts/commerce.js';
 
 // Initialize
 import '../../scripts/initializers/auth.js';
 import '../../scripts/initializers/order.js';
-import { rootLink } from '../../scripts/scripts.js';
 
 const renderSignIn = async (element, email, orderNumber) => authRenderer.render(SignIn, {
   initialEmailValue: email,
