+++ blocks/product-list-page-custom/product-list-page-custom.js
@@ -5,8 +5,7 @@ import htm from './htm.js';
 import ProductList from './ProductList.js';
 import FacetList from './FacetList.js';
 import { readBlockConfig, sampleRUM } from '../../scripts/aem.js';
-import { priceFieldsFragment, performCatalogServiceQuery } from '../../scripts/commerce.js';
-import { rootLink } from '../../scripts/scripts.js';
+import { priceFieldsFragment, performCatalogServiceQuery, rootLink } from '../../scripts/commerce.js';
 
 const html = htm.bind(h);
